+++ test/premature-exit.ts
@@ -2,5 +2,5 @@ import 'source-map-support/register'
 import { test } from '../index'
 
 test('kill -9 in test', () => {
-    process.kill(9)
+    process.kill(process.pid, 9)
 })